refactor: extract hotkey router from hook detours
